Fix bug detecting import or __all__ access failure
The left side of the pipe in the tox command, which is what ran the
per-environment python interpreter to import the module and access
its __all__ attribute, could fail, without causing the command as a
whole to fail (because the right side would still sucessfully sort
the zero lines of output from the left side).

This fixes that. It does so in a way that is bash-specific, but
the chronicall script itself is a bash script, and since its
hashbang is "#!/usr/bin/env bash", it will typically select the
same bash interpreter a tox environment will select with tox.ini
having "bash" in allowlist_externals.

Now the tox environment is correctly reported as failing when an
ImportError, or AttributeError on __all__ access, or any other such
error, occurs.